Fixing bugs
Some functions weren't working because  there was a comprison between ints and arrays, which is not well defined in python. To fix that I've changed the functions to show an error message whenever any value of the array satisfacted the comparison with the int, and to put a NaN in those points. The NaN's will have to be treated later so they won't iterfere in the functioning of other funtions.